requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
python-dotenv
openai
spacy
sacrebleu